+++ R/config.R
@@ -7,25 +7,26 @@
 #'
 #' @param dir Directory of the config file
 #' @param file Config file name
+#' @param default Default value for if a test should run (logical)
 #' @details If a test is missing from the config file, it is assume to be TRUE.
 #' Therefore, the config file can be quite short and just list exceptions.
 #' If the config file is missing, then all tests are carried out.
 #' @export
-create_config = function(dir = ".", file = "config-uat.yml") {
+create_config = function(dir = ".", file = "config-uat.yml", default = TRUE) {
   obj_info = get_check_info(dir, file)
   groups = unique(obj_info$group)
   shorts = purrr::map(groups, ~obj_info[obj_info$group == .x, ]$short)
-  group_shorts = purrr::map(shorts, create_group_short)
+  group_shorts = purrr::map(shorts, create_group_short, default = default)
 
   names(group_shorts) = groups
   yaml::write_yaml(group_shorts, file = file.path(dir, file))
   return(invisible(group_shorts))
 }
 
-create_group_short = function(short) {
+create_group_short = function(short, default) {
   group_short = vector("list", length = length(short))
   names(group_short) = short
-  purrr::map(group_short, ~TRUE)
+  purrr::map(group_short, ~default)
 }
 
 get_check_info = function(dir, file) {