+++ NEWS.md
@@ -1,3 +1,7 @@
+# jrHealthCheckConnect 0.2.12 _2022-12-20_
+- feat: Add plumber test
+- feat: Set config file with default "no"
+
 # jrHealthCheckConnect 0.2.11 _2022-12-20_
 - feat: Advance quarto tests: python/Julia/Obser'